1st KQBF
Changes:

Create function shouldInsert in Evaluator to see if  I should insert a value based on its weight. To QBF this is always true. To KQBF this is true If the weight is not exceeded.

Used shouldInsert in the default aka constructiveHeuristic() and localSearch(). In localSearch() if I can't insert the element I remove it from the candidate list anyways because I need to search another element to insert.